requirements.txt
Riskfolio-Lib==3.1.1
investpy==1.0.8
pandas==1.3.4
matplotlib==3.5.2
numpy==1.22.3
streamlit==1.9.0
